requirements.txt
nltk==3.8.1
openai==0.28.0
pandas==2.2.2
pdfplumber==0.11.0
python_docx==1.1.0
streamlit==1.32.0
textblob==0.18.0.post0
transformers==4.39.3
